Fix some NVM install issues
The `source $(brew --prefix nvm)/nvm.sh` in `nvm/install.sh` seems to be
required as otherwise the `nvm` command is not available in that script.

NVM versions 7 and 8 are not installed for the moment, as the Ember
versions with which we are working at the moment, don't compile with
node >= 7.

The default node version is set to 6, although it seems that this is not
automatically loaded when a new shell is opened.

A manual `nvm alias default 6` seems to do the trick.